new(vim): now checks for vim-plug installation and only installs if not found

+++ vim/install.sh
@@ -1,27 +1,35 @@
 #!/bin/bash
 
 if command -v vim >/dev/null 2>&1; then
-    echo "::[ Installing ] Vim >> 'vim-plug'"
-    curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
-        https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+    MODULE="Vim"
+    PLUG_FILE=~/.vim/autoload/plug.vim
 
-    echo "::[ Linking ] Vim >> '.vimrc'"
-    ln -sf "$DOTFILES/vim/vimrc.symlink" ~/.vimrc
+    if [ ! -f $PLUG_FILE ]; then
+        echo "[ INSTALLING ] :: $MODULE >> vim-plug"
+        curl -fLo $PLUG_FILE --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 
-    # vim +'PlugInstall --sync' +qa
-    vim +'PlugUpdate' +qa
+        echo "[ LINKING ] :: $MODULE >> .vimrc"
+        ln -sf "$DOTFILES/vim/vimrc.symlink" ~/.vimrc
+    fi
+
+    echo "[ UPDATING ] :: $MODULE >> Plugins"
+    vim +'PlugUpdate --sync' +qa
 fi
 
 if command -v nvim >/dev/null 2>&1; then
-    echo "::[ Installing ] Nvim >> 'vim-plug'"
-	curl -fLo ~/.local/share/nvim/site/autoload/plug.vim --create-dirs \
-		https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+    MODULE="NeoVim"
+    PLUG_FILE=~/.local/share/nvim/site/autoload/plug.vim
 
-	mkdir -p ~/.config/nvim/
+    if [ ! -f $PLUG_FILE ]; then
+        echo "[ INSTALLING ] :: $MODULE >> vim-plug"
+        curl -fLo $PLUG_FILE --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 
-    echo "::[ Linking ] Nvim >> 'init.vim'"
-    ln -sf "$DOTFILES/vim/vimrc.symlink" ~/.config/nvim/init.vim
+        mkdir -p ~/.config/nvim/
 
-    # nvim +'PlugInstall --sync' +qa
-	nvim +'PlugUpdate' +qa
+        echo "[ LINKING ] :: $MODULE >> init.vim"
+        ln -sf "$DOTFILES/vim/vimrc.symlink" ~/.config/nvim/init.vim
+    fi
+
+    echo "[ UPDATING ] :: $MODULE >> Plugins"
+	nvim +'PlugUpdate --sync' +qa
 fi
